Revert "Reland: "liblog: use a blocking socket for sending messages to logd"""
When readers connect, particularly when they use 'tail', they may hold
the logd lock for extended periods of time (>1s), causing the logd
socket to fill and writers to block.  This has resulted in various
bugs.  The 'tail' code and a majority of the log reading code in logd
needs to be refactored to properly fix this, so this must be reverted
in the meantime.

This reverts commit a65d14f82041898096a4e2a3c317ee74921826a9.

Bug: 164973960
Test: no log writes block
Change-Id: I265b961e0fa4fd10dc0118c291c20756f75dfc11