logd: Fix unintialized memory assumptions part 2
A previously change fixed that SerializedLogChunk.three_logs assumes
that log buffers are zero-initialized, but they are not.

Unfortunately, that fix wasn't enough since
sizeof(expected_buffer_data) includes the null terminator of the
expected string, but that null terminator is not expected to appear in
the log chunk.

This change subtracts 1 from sizeof(expected_buffer_data) to fix this
test on host (again).

Test: logd unit tests
Change-Id: Ibe824329116b1592df13b4069e5b000ee5cdb8bf