Add logd.ready
logd.ready is a system property that logd sets when it is ready to
serve incoming socket requests for reading and writing log. Clients of
logd (e.g. logcat) can use this to synchornize with logd, otherwise they
may experience a crash due to the refused socket connection to logd. The
property is automatically unset when logd is not running (e.g. being
restarted due to a crash, etc.).

Bug: 206826522
Test: run microdroid. see logcat logs are shown immediately.
Change-Id: I4397fcfd1b5dfe478005f5956a150ec94705ff14